# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/chats": {
      "post": {
        "operationId": "chat_createChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 생성",
        "tags": [
          "chats"
        ]
      },
      "get": {
        "operationId": "chat_search",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "없으면 처음부터",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "검색할 사용자 이름",
            "schema": {
              "example": "user123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 목록 조회",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/batch-delete": {
      "post": {
        "operationId": "chat_deleteManyChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteChatsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 여러개 삭제",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{id}": {
      "delete": {
        "operationId": "chat_deleteChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "채팅방이없거나 내가 그방 유저가 아니거나",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 삭제",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{id}/user": {
      "get": {
        "operationId": "chat_getOpponentUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpponentUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상대 유저 조회",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{id}/join": {
      "put": {
        "operationId": "chat_joinChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinChatRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 입장",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{id}/leave": {
      "put": {
        "operationId": "chat_leaveChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveChatRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 나가기",
        "tags": [
          "chats"
        ]
      }
    },
    "/chat-messages": {
      "post": {
        "operationId": "chatMessage_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatMessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "내가 그 채팅방의 일원이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "채팅방이 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅 보내기",
        "tags": [
          "chat-messages"
        ]
      },
      "get": {
        "operationId": "chatMessage_getMessages",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "없으면 처음부터",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "query",
            "description": "채팅방 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessagesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅방 메세지 조회",
        "tags": [
          "chat-messages"
        ]
      }
    },
    "/chat-messages/{id}/like": {
      "put": {
        "operationId": "chatMessage_likeMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "메시지가없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅 좋아요",
        "tags": [
          "chat-messages"
        ]
      },
      "delete": {
        "operationId": "chatMessage_unlikeMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "메시지가없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "채팅 좋아요 취소",
        "tags": [
          "chat-messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "grimity API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserBaseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "example": "profile/{UUID}.jpg",
            "nullable": true
          },
          "url": {
            "type": "string",
            "description": "라우팅용 url"
          }
        },
        "required": [
          "id",
          "name",
          "image",
          "url"
        ]
      },
...
      "IdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateChatRequest": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string"
          }
        },
        "required": [
          "targetUserId"
        ]
      },
      "LastChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true,
            "deprecated": true,
            "description": "DEPRECATED: images[0]와 동일. images 사용 권장",
            "example": "https://image.grimity.com/chat/123.webp"
          },
          "images": {
            "description": "메시지에 묶인 이미지들 (최대 5개, 없으면 빈 배열)",
            "example": [
              "https://image.grimity.com/chat/123.webp"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "senderId": {
            "type": "string",
            "description": "보낸 사람의 ID (마지막 채팅을 내가 보냈을수도잇음)"
          }
        },
        "required": [
          "id",
          "content",
          "image",
          "images",
          "createdAt",
          "senderId"
        ]
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "unreadCount": {
            "type": "number"
          },
          "enteredAt": {
            "format": "date-time",
            "type": "string"
          },
          "opponentUser": {
            "description": "채팅 상대의 사용자 정보",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserBaseResponse"
              }
            ]
          },
          "lastMessage": {
            "nullable": true,
            "description": "채팅방의 마지막 메시지",
            "allOf": [
              {
                "$ref": "#/components/schemas/LastChatMessageResponse"
              }
            ]
          }
        },
        "required": [
          "id",
          "unreadCount",
          "enteredAt",
          "opponentUser",
          "lastMessage"
        ]
      },
      "ChatsResponse": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "chats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatResponse"
            }
          }
        },
        "required": [
          "nextCursor",
          "chats"
        ]
      },
      "BatchDeleteChatsRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "OpponentUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "example": "profile/{UUID}.jpg",
            "nullable": true
          },
          "url": {
            "type": "string",
            "description": "라우팅용 url"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "isBlocking": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "image",
          "url",
          "isBlocked",
          "isBlocking"
        ]
      },
      "JoinChatRequest": {
        "type": "object",
        "properties": {
          "socketId": {
            "type": "string"
          }
        },
        "required": [
          "socketId"
        ]
      },
      "LeaveChatRequest": {
        "type": "object",
        "properties": {
          "socketId": {
            "type": "string"
          }
        },
        "required": [
          "socketId"
        ]
      },
      "CreateChatMessageRequest": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "description": "본문내용",
            "nullable": true
          },
          "images": {
            "example": [
              "v2/chat/{UUID}.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 5,
              "minLength": 0
            }
          },
          "replyToId": {
            "type": "string",
            "description": "답장 채팅메시지 ID",
            "nullable": true
          }
        },
        "required": [
          "chatId",
          "images"
        ]
      },
      "ReplyToResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true,
            "deprecated": true,
            "description": "DEPRECATED: images[0]와 동일. images 사용 권장",
            "example": "https://image.grimity.com/chat/123.webp"
          },
          "images": {
            "description": "메시지에 묶인 이미지들 (최대 5개, 없으면 빈 배열)",
            "example": [
              "https://image.grimity.com/chat/123.webp"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "image",
          "images",
          "createdAt"
        ]
      },
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true,
            "deprecated": true,
            "description": "DEPRECATED: images[0]와 동일. images 사용 권장",
            "example": "https://image.grimity.com/chat/123.webp"
          },
          "images": {
            "description": "메시지에 묶인 이미지들 (최대 5개, 없으면 빈 배열)",
            "example": [
              "https://image.grimity.com/chat/123.webp"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserBaseResponse"
          },
          "isLike": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "USER",
              "COMMISSION_REQUESTED",
              "COMMISSION_ACCEPTED",
              "COMMISSION_REJECTED",
              "COMMISSION_CANCELED",
              "COMMISSION_RESULT_UPLOADED",
              "COMMISSION_FINAL_UPLOADED",
              "COMMISSION_COMPLETED"
            ],
            "description": "메시지 타입 (USER: 일반, COMMISSION_*: 시스템 메시지)"
          },
          "referenceId": {
            "type": "string",
            "nullable": true,
            "description": "시스템 메시지 클릭 시 이동 대상 ID (커미션 등). 일반 메시지는 null"
          },
          "replyTo": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReplyToResponse"
              }
            ]
          }
        },
        "required": [
          "id",
          "content",
          "image",
          "images",
          "createdAt",
          "user",
          "isLike",
          "type",
          "referenceId",
          "replyTo"
        ]
      },
      "ChatMessagesResponse": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            }
          }
        },
        "required": [
          "nextCursor",
          "messages"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "유효성 검사 실패 시 문자열 배열로 내려온다"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}